Prepare to generalize logs
