+++ bin/generic/tag/add/examples.md
