+++ bin/generic/tag/delete/examples.md
