+++ bin/vm/index.js
@@ -10,7 +10,7 @@ const resource = {
     defaultQuery: '[].{id:_id,name:name,flavour:flavour,state:state,processing:processing,tags:join(\',\',keys(tag || `{}`) ) }',
     url: () => 'vm',
     plugins: genericDefaults.plugins,
-    commands: [ 'list', 'show', 'history'],
+    commands: [ 'list', 'show', 'history', 'tag'],
     title: 'virtual machine',
 };
 
@@ -48,7 +48,6 @@ category.addChild(require('./action/userdata')(actionDefault));
 category.addChild(require('./disk'));
 category.addChild(require('./nic'));
 category.addChild(require('./dvd'));
-category.addChild(require('./tag'));
 
 category.addChild(require('./ssh/ssh')(childDefaults));
 category.addChild(require('./serialport')(resource));
